fix(weave): prevent manual steps from being skipped on resume (#1372)
Mark `tool = "manual"` steps with `manual-handoff` status instead of
`completed` in the journal. On resume, re-display manual step instructions
rather than silently skipping them. This ensures mktsk and other manual
handoff steps are never lost when a workflow is resumed.
