+++ crates/cli-sub-agent/src/plan_cmd_tests_tail.rs
@@ -159,68 +159,6 @@ fn load_pr_bot_step_by_title(title: &str) -> PlanStep {
 }
 
 #[tokio::test]
-async fn execute_plan_stops_after_manual_handoff() {
-    let plan = ExecutionPlan {
-        name: "manual-handoff".into(),
-        description: String::new(),
-        variables: vec![],
-        steps: vec![
-            PlanStep {
-                id: 1,
-                title: "manual-step".into(),
-                tool: Some("manual".into()),
-                prompt: "Use mktsk in the main agent, then resume.".into(),
-                tier: None,
-                depends_on: vec![],
-                on_fail: FailAction::Abort,
-                condition: None,
-                loop_var: None,
-                session: None,
-            },
-            PlanStep {
-                id: 2,
-                title: "should-not-run".into(),
-                tool: Some("bash".into()),
-                prompt: "```bash\ntouch should-not-run\n```".into(),
-                tier: None,
-                depends_on: vec![],
-                on_fail: FailAction::Abort,
-                condition: None,
-                loop_var: None,
-                session: None,
-            },
-        ],
-    };
-    let vars = HashMap::new();
-    let tmp = tempfile::tempdir().unwrap();
-    let workflow_path = tmp.path().join("workflow.toml");
-    let journal_path = tmp.path().join("manual-handoff.journal.json");
-    std::fs::write(&workflow_path, "[workflow]\nname='manual-handoff'\n").unwrap();
-    let completed = std::collections::HashSet::new();
-    let mut journal = PlanRunJournal::new("manual-handoff", &workflow_path, vars.clone());
-    let mut run_ctx = PlanRunContext {
-        project_root: tmp.path(),
-        workflow_path: &workflow_path,
-        config: None,
-        tool_override: None,
-        journal: &mut journal,
-        journal_path: Some(&journal_path),
-        resume_completed_steps: &completed,
-        chunked: false,
-        no_fs_sandbox: false,
-    };
-
-    let results = execute_plan_with_journal(&plan, &vars, &mut run_ctx)
-        .await
-        .expect("manual handoff plan should execute");
-
-    assert_eq!(results.len(), 1, "manual handoff must pause the workflow");
-    assert_eq!(journal.status, "manual-handoff");
-    assert!(journal.completed_steps.contains(&1));
-    assert!(!tmp.path().join("should-not-run").exists());
-}
-
-#[tokio::test]
 async fn execute_plan_stops_for_await_user() {
     let plan = ExecutionPlan {
         name: "await-user".into(),
